used Int16 rather than Int32 for the weight in the AccountKeyAbi

+++ eos-swift/eos-chain-actions/transaction/account/actions/newaccount/AccountKeyAbi.swift
@@ -2,5 +2,5 @@ import Foundation
 
 struct AccountKeyAbi : Encodable {
     let key: PublicKeyWriterValue
-    let weight: Int32
+    let weight: Int16
 }
